Rename CounterIsPlacedTrigger to OneOrMoreCounterIsPlacedTrigger
