+++ src/magic/model/MagicAbility.java
@@ -1522,7 +1522,7 @@ public enum MagicAbility {
     WhenOneOrMoreCountersArePlacedOnSN("Whenever (you put)? one or more (?<type>[^ ]+) counter(s)? (are placed)? on SN," + ARG.EFFECT, 0) {
         @Override
         protected void addAbilityImpl(final MagicAbilityStore card, final Matcher arg) {
-            card.add(CounterIsPlacedTrigger.createOneOrMoreSelf(
+            card.add(OneOrMoreCounterIsPlacedTrigger.createSelf(
                 MagicCounterType.getCounterRaw(arg.group("type")),
                 MagicRuleEventAction.create(ARG.effect(arg))
             ));
